Move configuration loading
